Extract Phase 61 record validators
